pyproject.toml
[project]
name = "pythondev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "pip>=25.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "python-dateutil>=2.9.0.post0",
]

[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 88
count = true
